refactor to separate solar orbit module
太陽軌道に関わる処理が肥大化したため分離
